requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.9.0
django-queryset-csv==1.1.0
djangorestframework==3.12.4
Pillow==8.3.2
python-dotenv==0.19.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
unicodecsv==0.14.1
